Fix for minor race condition when making bug reports (https://github.com/runtimeverification/pyk/pull/637)
In https://github.com/runtimeverification/evm-semantics/pull/2047, we
have a minor race condition when creating bug reports where one thread
will check if the directory to create the report in exists, see that it
doesn't and then create it. Meanwhile, another thread will already have
created the directory, so it the first thread will fail because now the
directory does exist. This allows for that case.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>