Add support for vacuous nodes (#587)
With https://github.com/runtimeverification/haskell-backend/pull/3637,
kore-rpc returns a new response with a `vacuous` stop reason and the
current state before rewriting the state to `#Bottom`.

Changes in this PR:
- added a new set `vacuous` to the kcfg. The reason to add the set here
and not in the `APRProof` is that the response comes from the backend,
and is not semantics related.
- modified `cterm_execute` to return if the result is vacuous.
- modified `extend` to add a `vacuous` node to the kcfg if
`cterm_execute` gets a `vacuous` response.
- modified `APRProof` and `APRBMCProof` to prevent marking `vacuous`
nodes as `pending` or `failing`.
- modified proof summaries to include `vacuous`.

<img width="827" alt="Screenshot 2023-08-08 at 16 42 48"
src="https://github.com/runtimeverification/pyk/assets/16517508/5e20228e-03e0-412f-8fb1-c5f9ac340ad9">

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Petar Maksimovic <petar.maksimovic@runtimeverification.com>
Co-authored-by: Petar Maksimović <PetarMax@users.noreply.github.com>