Add K-ST (#3631)
We develop K-ST, a formal executable semantics for ST in the K
framework.
Defined with respect to the IEC 61131-3 standard and PLC vendor manuals,
K-ST is a high-level reference semantics that can be used to evaluate
the correctness and consistency of different ST implementations.

Detailed information is in https://arxiv.org/abs/2202.04076.