Adding  CI to release workflow to create a `develop` baseline performance on Bencher (#3627)
This modification is necessary to create a baseline for newer branches
to compare the performance of their changes against what we already have
in develop.

This modification is necessary to create and populate the develop's
branch as the baseline for newer branches.
This script `post_results_to_develop.py` works by getting the last
benchmarks reports of the last branch merged into develop, aggregating
them in one file, and posting them on the develop branch of Bencher.

Hopefully, this is a temporary fix, as this was already accepted as a
bug/possible feature by Bencher's maintainer
https://github.com/bencherdev/bencher/issues/187.

This PR, unlike https://github.com/runtimeverification/k/pull/3625,
doesn't need to rerun all `Performance Tests` on the develop or master
branch.