Ensure that requires clauses are included in KCFG target states (#648)
I was investigating a proof that is working with the legacy prover on
KEVM but failing on the RPC prover, specifically [this
one](https://github.com/runtimeverification/evm-semantics/blob/master/tests/specs/erc20/ds/transferFrom-failure-2-c-spec.k).

This proof was making it to the end of symbolic execution with the RPC
prover very quickly (~ 3 minutes), and then spending half an hour on the
implication check and eventually failing it. I minimized the example as
much as possible, and included it as a test in this PR. It turns out, it
cannot complete the implication check because the constraints from the
requires clause are not also included in the target state here.

This PR adds this test, which also times out after a long time spent
trying to prove the final implication (before these changes), then adds
the fix (which is just to include the requires clause constraints in the
target state of KCFGs too). Then the test passes in 19s (including
kompilation time).

---------

Co-authored-by: devops <devops@runtimeverification.com>