Store failure info in `Proof` (https://github.com/runtimeverification/pyk/pull/654)
I'm trying to simplify the kontrol code that computes failure logs for
each of the failing proofs, and I thought maybe it would be simpler if
we could just pass around the Proof object and have any necessary info
be available from there. We already have functions of `Proof` that
pertain to the proof status. Currently in KEVM to get the failure log,
we have to call a function separately when we detect a proof has failed
that computes this log, which requires a `KCFGExplore`, and also takes a
nontrivial amount of time. We then have to pass this around as part of a
tuple. This makes the computing of the failure info automatically happen
when a failing proof is detected, and saves the `APRFailureInfo` object
as a field in the `APRProof`.

The potential drawback I see to this is it would hurt performance on
failing proofs if the user doesn't care about generating the
`APRFailureInfo` for the proof.

---------

Co-authored-by: devops <devops@runtimeverification.com>