Fixing `develop.yml` workflow (#3655)
This PR fixes the develop workflow by:
- Modifying the `post_results_to_develop.py` script to get only the
branch's name without the user and adding one more logging message.
- Modifying the `develop.yml` to set the `${COMMIT_SHA}` env variable
and use it as a parameter to the above script, so we don't have to use
the `git` command inside the container raising ownership issues and
fixing a mismatch of `'` and `"` on the `bencher run` posting command.