Ordering instance for Sentences (2) (#3670)
fixes: #3530 

This branch was merged and then reverted because of an issue that didn't
get caught earlier in CI. I have since pushed a fix for the issue and
tested it to make sure that it will work. So, this should be good to go.
But if we really want to be sure, then we can consider this PR blocked
on #3669.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>