Updating `Guice` to 4.0 and deleting temporary injections (#3644)
This modification aimed to update the
[Guice](https://github.com/google/guice) version from `4.0-beta5` to
`4.0`, deleting the constructors with one void parameter to
```
Prevents instantiation by Guice when not explicitly configured in a Module.
```
And applying the `requireAtInjectOnConstructors` on every
`constructor()` inside a class that extends the `AbstractModule` for
consistency. These modifications made it necessary to explicitly create
constructors with zero parameters and within the `@Injector` annotation.

---------

Co-authored-by: devops <devops@runtimeverification.com>