Use prebuilt z3 images (#3694)
This uses the prebuilt Z3 images from:
https://github.com/runtimeverification/z3-images

This should speed up CI process.

Note that the debian packagaes are not changed in Z3 image used, just
the normal CI. So we are still testing with version 4.8.15 as well in
the packaging process. We could also change that to 4.12.1, but then we
wouldn't be testing the minimum recommended version.

Co-authored-by: Freeman <105403280+F-WRunTime@users.noreply.github.com>