Rephrase `sed` invocation for macOS (#3724)
This PR fixes a really tiny papercut in the regression test suite; out
of the box on macOS, the version of `sed` that's installed [doesn't
understand the `+` regex
operator](https://stackoverflow.com/questions/4453760/how-to-escape-plus-sign-on-mac-os-x-bsd-sed),
which causes one test to fail by not slicing out a non-deterministic
part of some test output.

Rather than trying to install a better version of `sed` for this one
case, we just rephrase the regex in question to use `*` instead.