Add option to print better parse errors (#3700)
This PR fixes #3672 by providing a best-effort rendering of the partial
state of the parser when a parse error happens. To do this, we grab the
Earley state corresponding to the span `[0, token_index_before_error]`
and inspect the partial parse trees available, as well as the production
being attempted at this point.

Consider as a simple example the following standard K definition for
arithmetic expressions:
```k
module ARITHMETIC-SYNTAX
  imports UNSIGNED-INT-SYNTAX
  imports ID-SYNTAX

  syntax Exp ::= Int
               | "-" Exp      [group(neg), strict]
               | Exp "+" Exp  [left, group(add), strict]
               | Exp "-" Exp  [left, group(sub), strict]
               | Exp "*" Exp  [left, group(mul), strict]
               | Exp "/" Exp  [left, group(div), strict]
               | "(" Exp ")"  [bracket]

  syntax priorities neg > mul div > add sub
endmodule

module ARITHMETIC
  imports ARITHMETIC-SYNTAX
endmodule
```

On trying to `kast` an invalid term (for example, `1 * 2 + a * 4` - `a`
lexes because we've imported `ID-SYNTAX`, but terms of sort `Id` can't
appear in a valid derivation of `Exp`), we see the following message
from the new component:
```console
$ kast <(echo '1 * 2 + a * 4') --no-exc-wrap
[Error] Inner Parser: Parse error: unexpected token 'a' following token '+'. Additional parsing diagnostic information:
  Attempting to apply production:
    syntax Exp ::= Exp "+" Exp
    produced partial term:
      `_+__ARITHMETIC-SYNTAX_Exp_Exp_Exp`(`_*__ARITHMETIC-SYNTAX_Exp_Exp_Exp`(#token("1","Int"),#token("2","Int")),#token("<error>","<invalid>"))

	Source(/dev/fd/11)
	Location(1,9,1,10)
	1 |
	  .	        ^
```
we make a best-effort attempt to construct what the top-level `_+_` term
might have looked like, placing an `<error>` token where we expected to
see a valid term of sort `Exp`.

The changes made in this PR are as follows:
* Update the implementation of `EarleyParser` to conditionally perform
the rendering as above.
* Add boilerplate code to thread a new option `--debug-parse` through to
the parser when appropriate.
* Add a test case based on the above arithmetic module.

I have additionally verified with @dkcumming that the implementation as
reviewed produces useful information that would have been useful when
addressing the original problem; these cases come from the MIR semantics
and are therefore too large to include here as tests.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>