KoreClient: add optional name for rpc directories in bug report (#693)
This change introduces an optional string to use as an ID in the bug
report archive. The string will become part of the `rpc_*` directory
that holds request and response JSON if `--bug-report` is used.

Previously, the names for `rpc_*` directories were using the object ID
of the `KoreClient` instance. This is the default if no custom ID is
provided.

### Background
In the most common use case (`evm-semantics` proofs), each claim in a
spec file is run by the pyk prover as a separate proof, including the
creation of a fresh `KoreClient`. However, all proofs will use the same
`definition.kore` and the same bug report archive - JSON data for the
different proofs is stored in different `rpc_*` directories in this
archive.
(The server is apparently also restarted for each proof but all
invocations use identical parameters)

---------

Co-authored-by: devops <devops@runtimeverification.com>