Fix for macOS nix / kup breakage (#3741)
On my Intel mac machine, the minimal JRE package is broken:
```console
$ /nix/store/6bqyc34isvgrrgl3nrqilzqbx7r1dyq5-zulu19.30.11-ca-jdk-minimal-jre-19.0.1/bin/java
[1]    31710 killed
```

This causes the default `kup`-installed K to break; the fix is just to
identify this platform and package the full JRE. The only downside here
is a slightly bigger closure for users on this platform, which is likely
to be deprecated at some point in the future anyway.

I have also added a new minimal "smoke test" to the Nix CI tests;
previously we were not checking that `kompile` etc. actually run on
Intel macOS.