Fix release by installing Calibre 6 (#3742)
This PR is a repeat fix for
https://github.com/runtimeverification/k/issues/3710; I realise that the
issue has recurred because the 5.x releases of Calibre are being
invalidated faster as they are "out of date". To address this, I tested
the relevant part of the release job with Calibre 6.29 in the PR CI
here; it all appears to work fine.