Pin nixpkgs to haskell-backend (#3754)
We are having issues with downstream packages overriding nixpkgs version
of upstream packages. We should try to invert this and instead pick up
nixpkgs from the upstream dependencies. Since haskell-backend is the
deepest dependency and also the most sensitive to nixpkgs version, i
have picked it to be the source of pinned nixpkgs for everything
downstream.