Apply "silent code cleanup" from IntelliJ (#3755)
This PR applies all automatic cleanups that IntelliJ is confident enough
to apply silently with no further user input; a rough summary of the
different kinds of change are:
* Removal of redundant casts (`T t = (T) v` where `v` already has type
`T`).
* Removal of unnecessary calls to `toString()`
* Insertion of `final` where possible
* Collapsing of redundant boolean conditions `if (cond) return true else
return false`
* Syntactic fixes (`;;` etc.)
* Use of `StandardCharsets.UTF_8` to simplify exception checks when
encoding strings
* Removal of redundant explicit key checks when removing values from
hashmaps

As ever, I'm happy to bikeshed these changes - I can redo the analysis
and disable some subset of the checks if anyone has objections to the
specific results we get.

~~To avoid merge conflicts, this PR is based on #3746 and should only be
merged when that PR is merged.~~

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>