Move mac release CI to hosted ARM runners (#3764)
This PR moves our macOS packaging and release workflow over to use our
self-hosted runners, which will hopefully address the recurrent
flakiness we've had due to resource exhaustion (#3705).

I have tested these changes by adapting the release workflow to run on
PRs, confirming that an `arm64_ventura` bottle gets successfully built.
I don't anticipate any changes to the orchestration logic being
required, but I will keep an eye on the release that this PR generates
just to make sure that it doesn't break the release worse than it
already is. In that event, this PR can easily be reverted or hotfixed.