`pyk rpc-kast` --- convert responses to new requests (https://github.com/runtimeverification/pyk/pull/700)
This PR extends `pyk` with `pyk rpc-kast` --- and utility to convert
Kore JSON RPC responses to new requests. Thiss functionality is useful
when doing rpc server-level debugging.

Supported conversions:
* "execute" response to "execute" request
* "execute" response to "simplify" request

The tool also expects a reference request, which will be used to detect
the type of the new request and fill-in the parameters. This way we can
avoid a ton of command-line arguments to the tool that would st various
request parameters. A good reference request is almost always around,
and the parameters could be fine-tuned manually or with tools like `jq`.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>