Temporary restore of arch PKGBUILD (#3770)
This PR fixes an issue with the master-branch version bumper introduced
in #3765; that PR removed the Arch package definition at the same time
as removing it from the version bump script. However, the master branch
job still has the old version of the version bumper, which introduces a
conflict when trying to modify the package definition.

The fix is to restore the old version of the package definition, allow
one master branch bump to go through, then follow up by removing the
package definition once the master branch isn't trying to modify it.