Revert experimental Homebrew changes (#3771)
This PR reverts a series of experimental changes I made last week to
attempt a fix for the flakiness of the macOS Homebrew job; the required
implementation turned out to be substantially trickier than I'd thought
and so this PR undoes the changes in the interest of returning the
release job to a better state. The PRs reverted are:
* https://github.com/runtimeverification/k/pull/3764
* https://github.com/runtimeverification/k/pull/3766
* https://github.com/runtimeverification/k/pull/3767
* https://github.com/runtimeverification/k/pull/3768

Additionally, I have temporarily disabled the PL-tutorial tests on Intel
macOS to avoid a common source of release flakiness.

Future changes to the Homebrew code should review and adapt the changes
made and reverted here.