Fix homebrew repo authentication (#3781)
The previous version of this workflow was doing something non-standard
with SSH keys that broke when we migrated it over to the RV org; this PR
fixes the breakage by using our standard `JENKINS_GITHUB_PAT`
authentication instead.

I have tested that this works by running the steps in isolation on a PR
branch (https://github.com/runtimeverification/k/pull/3780) and I'm
reasonably confident this fix is correct.