Bump LLVM version in packaging to 15 (#3783)
In https://github.com/runtimeverification/llvm-backend/pull/874 we
bumped the minimum version of LLVM required by the LLVM backend to 15;
we can safely do this now as we no longer need to support Ubuntu Focal.
This PR is the corresponding update to K; it updates any remaining
references to LLVM 14 to LLVM 15.

~~Can be reviewed, but don't merge until
https://github.com/runtimeverification/homebrew-k/pull/2 is merged~~

Will unblock: https://github.com/runtimeverification/k/pull/3782