Remove mavenix take 2 (#3806)
Switch from using mavenix to using `buildMavenPackage` back-ported from
the upcoming nixpkgs 23.11 stable release. This may increase build times
somewhat and updating maven dependencies will no longer be automatic,
but arguably the automation never works and It's difficult to figure out
why it's failing.

I've updated the README troubleshooting section with the two cases of
manual intervention, necessary with the new packaging solution, when
maven dependencies change.

The hash mismatch should only happen when deps change, but if we want to
automate this in ci then we can do something similar to:
https://github.com/runtimeverification/hs-backend-booster/blob/175f10b16ad50dd2bb50fcfe68cbcacd79760317/scripts/update-haskell-backend.sh#L21C3-L22