Add `total` to `#let`-bindings where LHS is a variable (#3798)
This PR addresses an issue whereby the Booster reverts to the old
Haskell backend when it encounters a rule written using `#let` bindings.
The reason for this is that the de-sugared form of `#let` is expressed
in terms of an intermediate `#lambda` function that is not marked as
being `total`.

To address the issue, this PR adds the `total` attribute to the
desugared intermediate lambda function for let-bindings and anonymous
functions when the left-hand side of the original binding is a variable.

Points to consider when reviewing this PR:
* ~~Is the identification of when the LHS a variable correct in its
handling of semantic casts and sorting? My presumption is that it is
given that the "first child of `#SemanticCastTo` `KApply`" logic is
replicated elsewhere.~~
*
https://github.com/runtimeverification/k/pull/3798#discussion_r1386805154
* Is there a way to integration-test this change other than "it doesn't
break the rest of the test suite"? At the moment I'm not sure there is
an obvious way to do so, but I'm open to suggestions.

Fixes https://github.com/runtimeverification/k/issues/3715

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>