Remove `EAsSet[E]` and `WAsSet[W]` traits in `Transformers.scala` to address IntelliJ error (#3840)
This PR removes the `EAsSet[E]` and `WAsSet[W]` traits in
`Transformers.scala`, and instead just duplicates a few lines of code
between `SetsGeneralTransformer` and `SetsTransformerWithErrors`.

As a result, we fix an annoying issue where IntelliJ incorrectly reports
an error on every inheritor of these classes, falsely claiming that
`errorUnit()` has an incorrect return type.

Given that the cost here is low, and similar issues have bitten us in
the past by blocking use of the debugger (#3816), I think this is a
worthwhile change. If no one is opposed, I plan to address any similar
incorrectly-reported IntelliJ errors as I encounter them.