Clean up `{...}<:S` strict cast syntax (#3853)
This PR cleans up a few issues related to the `{...}<:S` syntax for
strict casts. Specifically,

- Add a section to Lesson 1.11 explaining the need for the braced strict
cast syntax
- Change the klabel from `#InnerCast` to `#SyntacticCastBraced`
   - `#InnerCast` was a holdover from prior to #2352 
- Opinionatedly, change the syntax from `{...}<:S` to `{...}::S`
- Makes it more obvious this is just an alternative syntax for strict
casts `::S`
- To me, the syntax `<:S` is misleading in that it reads as "subsort of
`S`", but the actual meaning is "exactly the sort `S` and not a proper
subsort".
- Minimal fallout - there is only one rule in `rv-match` and two tests
in `pyk` where this needs to be changed
(https://github.com/search?q=org%3Aruntimeverification+%7D%3C%3A&type=code)

The commit history is clean and I can just drop the last commit if we
don't want to change the syntax.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>