Prohibit user-defined parametric sorts (#3878)
This PR adds a check to error on any non-`MInt` parametric sorts. 

We already assume `MInt` is the only parametric sort internally, so this
check just makes that explicit to the user:

https://github.com/runtimeverification/k/blob/0b4353c62f018a46736dc6b512414b77ab81523d/kernel/src/main/java/org/kframework/parser/inner/RuleGrammarGenerator.java#L419

The check has to be done in `Module.checkSorts` so that we error early
enough, e.g., before computing the subsort `POSet`.