Adjustments to allow Kontrol to generate usable KEVM specifications (https://github.com/runtimeverification/pyk/pull/804)
This PR enables multiple things used in generating KEVM claims from
Kontrol specifications. In particular:

- The `geInt` helper is added to the Kast prelude.
- Allow user to pretty-print in a specified module of the given
definition in `KPrint.pretty_print`.
- The branch recognizer which tries to extract branch conditions from
returned states from the backend is a bit smarter about branches due to
integer disequalities (a common case in KEVM proofs).
- Two new options are added to the reachability prover which allow
controlling how fast/often the subsumption check occurs:.
- `always_check_subsumption` allows disabling subsumption check on nodes
that are not marked as terminal, to save time.
- `fast_check_subsumption` allows using a much faster in-pyk check for
subsumption before consulting the backend. The faster check extracts the
K cell and only checks it for subsumption. If this check fails, then the
slower backend check is skipped for the given node (the old backend is
still consulted on success). This check is also used when marking nodes
as terminal, if something "looks" terminal in the K-cell, then it is
considered terminal with this check (this is so that when used in
conjunction with `always_check_subsumption`, we make sure to actually
check these nodes too).

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>