Log rewrites on edges in KCFG (https://github.com/runtimeverification/pyk/pull/817)
In order to better debug what is happening during symbolic execution of
basic blocks, this starts attaching some debug information to the
`KCFG.Edge` data-structure regarding the rewrites that happened for that
chunk of execution. In particular:

- The `KCFG.Edge` now gets a `rules` field, which stores the list of
rule labels that applied to make that edge.
- The `KSentence`, `KFlatModule`, and `KDefinition` classes are extended
with functionality to extract more human-readable labels and sources
from rule unique ids (supplied back by the backend).
- The `KCFGExplore.extend_kcfg` is modified to make sure that we attach
this extra information to `KCFG.Edge` when creating it.
- Fix an issue in the `KCFGViewer` where it wasn't actually able to call
the custom viewer on any `KCFG.Successor` because it was doing an exact
class comparison instead of subclass (using `isinstance`).

Note that this isn't quite as useful until
https://github.com/runtimeverification/hs-backend-booster/issues/466 is
merged, but still is working when the backend is reporting back rule
lists correctly.

---------

Co-authored-by: devops <devops@runtimeverification.com>