Attach logs to NDBranch, small refactors (https://github.com/runtimeverification/pyk/pull/831)
These are small changes that were made for setting up CSE on Kontrol. In
particular:

- `NDBranch` are setup to display their rule applications. A helper is
factored out for doing this for both edges and ndbranches.
- An function-internal variable is prefixed with `_` to avoid shadowing
confusing with outer variable.

The rule labels attached to `NDBranch` are not populated yet, due to
https://github.com/runtimeverification/hs-backend-booster/issues/472,
but once that bug is fixed then this will become more useful.

---------

Co-authored-by: devops <devops@runtimeverification.com>