add new classpath dir in ~/.local (#3927)
This adds a new directory to search for jars in the K frontend. This
directory is located in ~/.local. This allows the user to install
extensions to K without modifying their global installation. This is
useful if you want to install the Maude backend on a Nix-installed K
installation, for example.