Fully qualify claim labels/dependencies, include dependencies in output (https://github.com/runtimeverification/pyk/pull/842)
We were having trouble in
https://github.com/runtimeverification/evm-semantics/pull/2272 selecting
the correct claims because they can be unqualified by the user. This PR:

- Fully qualifies all claim labels and dependency labels returned by
`KProve.get_claim_modules`.
- Makes sure that we return not only the user-selected claims, but their
dependencies (unless the user turns that off), from `KProve.get_claims`.
- Adds a test-suite for the claim selection functionality.

There is a bug in this code, and a test added (and skipped using
`pytest.skip()`) which demonstrates a (probably minor) issue where we
are not checking that claim dependencies are in a module imported by the
current module.

---------

Co-authored-by: devops <devops@runtimeverification.com>