Base64-encode module digests (#3951)
Previously, the content generated for module digest attributes was an
arbitrary sequence of bytes. This broke outer parsing for `compiled.txt`
(i.e. the compiler is generating an intermediate state that it can't
re-compile). To fix the issue, we just apply a base 64 encoding to the
digest value when it's generated to keep the byte values in printable
ASCII range.

Fixes: https://github.com/runtimeverification/k/issues/3945