Fix for dotted attribute names in outer parser (#3953)
We serialize Java objects directly to represent the source and location
attributes in `compiled.txt`; this breaks the part of the outer parser
responsible for lexing an identifier name
(https://github.com/runtimeverification/k/issues/3952). To avoid having
to rewrite every bit of code that deals with these attributes, we
instead just make the lexical specification of attribute names a bit
more relaxed.

Fixes #3952

Co-authored-by: rv-jenkins <admin@runtimeverification.com>