Add one-argument version of `symbol(_)` (#3978)
This PR is the first change identified as necessary in #3966; it takes a
necessary but insufficient step towards rationalising the confusing
`klabel` and `symbol` attributes.

The change made here is to permit a new arity-1 version of the `symbol`
attribute, such that `symbol(X)` behaves identically to `klabel(X),
symbol`. Doing so will allow us to begin the process of deprecating the
`klabel` attribute once we have an alternative formulation for
overloaded symbols.

The code changes required are small; I have added a check to ensure that
new code doesn't mix the old and new idioms, as well as a test and the
beginning of new documentation for `symbol(_)` that can be updated as we
make related changes here.